Modified Masterscene + GP Env/Gameplay
Masterscene: created folder "BASE" which is unloaded. It has to stay in MS as reference, but it won't work properly in gameplay scene. In order to make it work, it has to be copied in Gameplay scene.
Folder BASE still need to be checked if it works in all its parts during gameplay. 
DIRECTIONAL LIGHT NEEDS TO STAY IN MASTERSCENE.